requirements.txt
py-clob-client
python-dotenv
pandas
gspread
gspread-dataframe
sortedcontainers
websockets==12.0.0
cryptography==42.0.8
web3==5.3.1
asyncio
requests
google-auth